requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
distro==1.9.0
exceptiongroup==1.2.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
instagrapi==2.1.2
openai==1.35.13
pillow==10.4.0
pycryptodomex==3.20.0
pydantic==2.7.1
pydantic-core==2.18.2
PySocks==1.7.1
requests==2.32.3
sniffio==1.3.1
tqdm==4.66.4
typing-extensions==4.12.2
urllib3==2.2.2
